requirements.txt
pyTelegramBotAPI>=4.15.2
pydantic>=1.10
openai>=1.12.0
apscheduler>=3.10
matplotlib>=3.7
python-dotenv>=1.0.0
